requirements.txt
Django>=4.2,<4.3
django-htmx>=1.18.0,<1.19.0
pillow>=10.4.0,<10.5.0
numpy>=2.1.0
opencv-python==4.10.*
pandas
plotly>=5.22.0,<5.23.0
tensorflow>=2.17.0
scikit-learn>=1.5.1,<1.6.0
django-background-tasks>=1.3.0,<1.4.0